[FLINK-16441][SQL] SQL Environment Configuration section will also be set in flink conf
